% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_concentation_homogeneity.R
\name{calc_concentation_homogeneity}
\alias{calc_concentation_homogeneity}
\title{Concentration and Homogeneity analysis for pool models}
\arguments{
\item{data_df}{('tbl') Dataframe containing all variables listed below.}

\item{lgd_pred_var}{('character') Name of the variable/column containing model predictions.}

\item{lgd_real_var}{('character') Name of the variable/column containing realised LGD.}

\item{grade_var}{('character') Name of the discrete variable indicating the grade/pools.}

\item{alpha}{('numeric') Confidence level 1-alpha/2 is diplayed in SUMMARY_TABLE}
}
\value{
tibble
}
\description{
This function performs a Concentration and Homogeneity analysis for pool models.
Outputs include a summary table depicting mean, standard deviation and confidence intervalls of the mean
("SUMMARY_TABLE"); the normalized Herfindahl Index ("HERFINDAHL_INDEX"); Boxplots per pool ("BOXPLOT");
Violin Plot per pool ("VIOLIN_PLOT") and a histogram depicting distribution across pools.
}
\examples{
conhom <- calc_concentation_homogeneity(data_df=df_lgd,lgd_pred_var=LGD_PREDICTION,
lgd_real_var=REAL_LGD,grade_var=LTV_BIN,alpha = 0.1)
}
