% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pairwise_wilcox.R
\name{calc_pairwise_wilcox}
\alias{calc_pairwise_wilcox}
\title{Paired Samples Wilcoxon Test for all pool pairs}
\arguments{
\item{data_df}{('tbl') Dataframe containing all variables listed below.}

\item{lgd_pred_var}{('character') Name of the variable/column containing model predictions.}

\item{lgd_real_var}{('character') Name of the variable/column containing realised LGD.}

\item{grade_var}{('character') Name of the discrete variable indicating the grade/pools.}
}
\value{
tibble
}
\description{
This function performs a Paired Samples Wilcoxon Test for the realized values of each pool pair and returns an matrix (tibble) of p-values.
Alternative Hypothesis: Pools have significantly different means.
}
\examples{
calc_pairwise_wilcox(data_df=df_lgd,lgd_pred_var=LGD_PREDICTION,
lgd_real_var=REAL_LGD,grade_var=LTV_BIN)
}
