% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_predictive_power_backtest.R
\name{calc_predictive_power_backtest}
\alias{calc_predictive_power_backtest}
\title{Predictive Power Test (T-Test on given level)}
\arguments{
\item{data_df}{('tbl') Dataframe containing all variables listed below.}

\item{lgd_pred_var}{(\code{character}) Name of the variable/column containing model predictions.}

\item{lgd_real_var}{(\code{character}) Name of the variable/column containing realized LGD.}

\item{grade_var}{(\code{character}) Name of the discrete variable indicating the grade/pools.}

\item{tol}{('numeric') tolerance level applied to t-test.}
}
\value{
tibble
}
\description{
This function is designed to compare the realization and predictions
of a continuous variable (LGD) on a given level (overall, pools or bins).
"P_CONS" (the p-value of the conservativity test) corresponds to the alternative hypothesis the real value is larger than the predicted value.
"P_CONS_TOL" corresponds to the same test as "P_CONS" but after shifting the prediction up by the tolerance level (e.g. 2.5\%).
"P_PROG" (the p-value of the progressiveness test) corresponds to the alternative hypothesis the real value is larger than the predicted value.
"P_PROG_TOL" corresponds to the same test as "P_PROG" but after shifting the prediction down by the tolerance level (e.g. 2.5\%).
}
\examples{
calc_predictive_power_backtest(df_lgd,lgd_pred_var=LGD_PREDICTION,lgd_real_var=REAL_LGD)
}
