% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_somersd_stats.R
\name{calc_somersd_stats}
\alias{calc_somersd_stats}
\title{Somers' D and Kendall's Tau B with conf. Intervalls}
\arguments{
\item{x}{('vector') Vector or trivial dataframe containing realised LGD.}

\item{y}{('vector') Vector or trivial dataframe containing predicted LGD.}

\item{conf_level}{('numeric') confidence level as double.}

\item{num_boot_samples}{('numeric') number of iterations used in bootstrap estimation of confidence bounds.}
}
\value{
list
}
\description{
This is a function which calculates the Somers' D and Kendall's Tau B statistic efficiently with bootstrapped confidence intervals. 500 bootstrap iterations are performed and the 80\% confidence level is used by default.
}
\examples{
calc_somersd_stats(x = c(0.1,0.2,0.3,0.4,0.5,0.6), y= c(0.2,0.2,0.2,0.2,0.3,0.3))
}
