% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAP_Curve.R
\name{CAP_Curve}
\alias{CAP_Curve}
\title{Cumulative Accuracy Profile Plot}
\arguments{
\item{sample}{('tbl') Dataframe containing variable and realised LGD.}

\item{var_target}{('character') Name of variable containing realised LGD.}

\item{var_pred}{('character') Name of variable containing predicted LGD or risk driver.}

\item{title}{('character') Title displayed above plot after "CAP Curve".}

\item{file}{('character') Folder path to save output image under the name given in title.}
}
\value{
CAP Curve image
}
\description{
This is a function which plots Cumulative Accuracy Profile for given realisations and estimates of a continuous variable.
}
\examples{
df <- tibble::tibble(x = c(0.1,0.2,0.3,0.4,0.5,0.6), y= c(0.2,0.2,0.2,0.2,0.3,0.3))
CAP_Curve(df, y, x, title="Toets")
}
