% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_custom_color_tile_backtesting.R
\name{fct_custom_color_tile_backtesting}
\alias{fct_custom_color_tile_backtesting}
\title{Format backtesting matrix}
\usage{
fct_custom_color_tile_backtesting(...)
}
\arguments{
\item{...}{('numeric')}
}
\value{
format
}
\description{
Function defines format for matrix with color scale.
}
