% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_format_perc_matric.R
\name{fct_format_perc_matrix}
\alias{fct_format_perc_matrix}
\title{Apply format to backtesting matrix}
\arguments{
\item{df}{('tbl')}

\item{title}{('character')}
}
\value{
kable
}
\description{
Function applies color scale format to matrixes with percent entries.
}
