% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_rc_estimate.R
\name{fct_rc_estimate}
\alias{fct_rc_estimate}
\title{Retail Construction Estimation/Calibration function}
\arguments{
\item{data_df}{('tbl') Dataframe containing the realized LGD, the coll. quota, a variable with coll. quota buckets and a variable with season buckets. NO MISSING VALUES}

\item{var_vec}{('vector') Vector of the form c("SEASON_BU","COLL_QUOTA_BU"), i.e. first the season bucket variable thereafter the coll. quota bucket.}

\item{lgd_real_var}{('character') Name of the variable/column containing realised LGDs.}

\item{coll_quota_var}{('character') Name of the variable/column containing the collateralization quota.}
}
\value{
('tibble') The output table includes the LGD estimate as LGD_PRED
}
\description{
The function performs the Retail Construction LGD MOdel estimate/calibration, i.e. it performs an isotonic regression over the collateralization quota and further differentiates uncollaterlaized observations via SEASON buckets.
}
