% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_rc_grades.R
\name{fct_rc_grades}
\alias{fct_rc_grades}
\title{Retail Construction Missing Treatment and Grades}
\arguments{
\item{data_df}{('tbl') Dataframe containing a COLL_QUOTA variable and a SEASON_REL variable}

\item{coll_quota_var}{('character') Name of the variable/column containing the collateralization quota.}
}
\value{
('tibble') The output table includes the SEASON Buckets (SEASON_BU), COLL_QUOTA Buckets (COLL_QUOTA_BU), and the GRADES (GRADES) variable
}
\description{
The function adds the Retail Construction LGD Model's buckets and grades to a dataframe containing a COLL_QUOTA variable and a SEASON_REL variable
}
