% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_table_format.R
\name{fct_table_format}
\alias{fct_table_format}
\title{Format output tables in Markdown}
\arguments{
\item{df}{('tbl')}

\item{title}{('character')}
}
\value{
kable
}
\description{
Function formats output tables to be presented in a HTML Markdown.
}
